fix: Fixed picture slider not consistently showing the active slide user clicked on upon opening the modal
The bug I fixed was when the user clicks on an image from the photo gallery in the info card,
a picture slider modal should open up and the image slider track should already be positioned to show,
the user what image they have clicked on. Sometimes, the user clicks on a image and the picture slider
modal is opened, but they have to scroll extra to the left of right in order to see what image they actually clicked on.
The proper positioning of the image slider track upon opening the picture slider modal, worked when I clicked
on the first image or the nth image of a photo gallery within a info card, but there were some edge cases where
the proper positioning of the slider track was not working properly like when clicking on the last image in a
photo gallery within a info card.