fix: Moved align-items: center; to correct CSS selector for my info card pictures
It's a pain in the butt to have to rebuild all of my production code, because I
want to transfer a single change I've made in my source code. I've even had to
manually relink the CSS and Javascript for my anime reviews page, character gallery
page,and my story synopsis page, because Parcel doesn't build the CSS and Javascript
file paths properly.